More open abstraction, allowing netty easy extending
